+++ src/Acta.Postgres/Sql/Operations/Tags/GetTags.sql
@@ -1,5 +1,5 @@
 WITH target (scope_id, namespace_id) AS (
-    SELECT t.id::bigint, NULL::smallint
+    SELECT t.id::bigint, NULL::int
     FROM {{schema}}.tenants t
     WHERE @p_scope_code = 20 /* TagScopeCode.Tenant */ AND t.tenant_key = @p_lookup_name
     UNION ALL
