+++ src/Acta.Relational/Connections/OutboxSourceDialect.cs
@@ -56,7 +56,7 @@ internal abstract class OutboxSourceDialect : ISqlDialect
 
     public void BindRegisterJobDefinitions(
         DbCommand command,
-        short namespaceId,
+        int namespaceId,
         DateTime manifestGenerationUtc,
         IReadOnlyList<JobDefinitionRow> rows,
         string schema
